Move the backend module out of bindings
We want to use this for the scaffolding code as well, so it seems like
it should live outside of bindings.

Split up the module into several submodules in preparation of adding
more code to them.
