+++ uniffi_bindgen/src/bindings/kotlin/gen_kotlin/miscellany.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::bindings::backend::{CodeOracle, CodeType, Literal};
+use crate::backend::{CodeOracle, CodeType, Literal};
 use askama::Template;
 use paste::paste;
 use std::fmt;
