+++ uniffi_bindgen/src/lib.rs
@@ -106,6 +106,7 @@ use std::{
     process::Command,
 };
 
+pub mod backend;
 pub mod bindings;
 pub mod interface;
 pub mod scaffolding;
